feat(P_IMPLEMENTATION_SP_C1_GREEN): feat(green): C1 — move stranded MCP tests to tests/mcp_server/unit/config/, delete dead dirs
- tests/unit/config/test_c_loader_structural.py → test_c_loader_schema_structural.py
  (renamed: avoids conflict with existing test_c_loader_structural.py; parent depth corrected)
- tests/unit/config/test_c_settings_structural.py → tests/mcp_server/unit/config/
- tests/unit/ removed from git tracking (git rm)
- proof_of_concepts/ removed from git tracking
- temp/ removed from git tracking
- src/copilot_orchestration/ + tests/copilot_orchestration/: were never tracked
- 2864 passed, 0 failed; quality gates 5/5 (#388)
