+++ .phase-gate/state.json
@@ -38,5 +38,5 @@
     }
   ],
   "reconstructed": false,
-  "current_sub_phase": null
+  "current_sub_phase": "red"
 }
