--- proof_of_concepts/real_async_poc.py
@@ -1,398 +1,0 @@
-"""
-REAL Async Event Architecture - Proof of Concept
-
-This demonstrates the ACTUAL problem with CPU-bound work in async Python.
-
-Test Scenarios:
-1. SYNC (baseline): Show events MISSED during blocking CPU work
-2. ASYNC (naive): Show asyncio.sleep() DOESN'T help with CPU work
-3. ASYNC + ProcessPool: Show TRUE parallelism solves the problem
-
-Run:
-    python proof_of_concepts/real_async_poc.py
-
-Expected results:
-- Scenario 1 (SYNC): Events missed during CPU work
-- Scenario 2 (ASYNC naive): STILL misses events (asyncio.sleep ≠ CPU work)
-- Scenario 3 (ProcessPool): NO missed events (true parallelism)
-"""
-
-import asyncio
-import contextlib
-import multiprocessing
-import time
-from concurrent.futures import ProcessPoolExecutor
-from datetime import datetime
-
-# ============================================================================
-# CPU-Intensive Work (The REAL problem)
-# ============================================================================
-
-
-def cpu_intensive_work(event_id: int, duration_ms: int) -> dict:
-    """
-    REAL CPU-bound work that BLOCKS the thread.
-
-    This simulates signal detection, risk calculation, etc.
-    NO asyncio.sleep() tricks - this is actual CPU work.
-    """
-    start = time.perf_counter()
-
-    # Busy loop - actual CPU work that blocks
-    target_duration = duration_ms / 1000.0
-    result = 0
-
-    while (time.perf_counter() - start) < target_duration:
-        # Simulate mathematical calculations (e.g., indicators, signals)
-        for i in range(1000):
-            result += i**2
-
-    elapsed = time.perf_counter() - start
-
-    return {"event_id": event_id, "result": result, "processing_time_ms": elapsed * 1000}
-
-
-# ============================================================================
-# Event Source (Simulates External Events)
-# ============================================================================
-
-
-class EventSource:
-    """
-    Simulates external event source (e.g., exchange WebSocket).
-
-    Publishes events at regular intervals.
-    """
-
-    def __init__(self, interval_ms: int = 100) -> None:
-        self.interval_ms = interval_ms
-        self.events_published = 0
-        self.publish_log: list[dict] = []
-
-    def publish_event(self) -> dict:
-        """Publish single event."""
-        event = {
-            "event_id": self.events_published,
-            "timestamp": datetime.now(tz=datetime.UTC),
-            "data": f"Event_{self.events_published}",
-        }
-
-        self.events_published += 1
-        self.publish_log.append({"event_id": event["event_id"], "published_at": event["timestamp"]})
-
-        return event
-
-
-# ============================================================================
-# Scenario 1: SYNCHRONOUS (Baseline - Shows The Problem)
-# ============================================================================
-
-
-def scenario_1_synchronous():  # noqa: ANN201
-    """
-    Synchronous event processing.
-
-    Expected: Events MISSED during CPU work (blocking).
-    """
-    print("=" * 80)
-    print("SCENARIO 1: SYNCHRONOUS (Baseline)")
-    print("=" * 80)
-    print("Publishing events every 100ms")
-    print("Processing takes 500ms (CPU-bound work)")
-    print("Expected: Events missed during processing")
-    print()
-
-    source = EventSource(interval_ms=100)
-    processed_events = []
-
-    start_time = time.perf_counter()
-    last_publish = start_time
-
-    # Run for 3 seconds
-    while (time.perf_counter() - start_time) < 3.0:
-        current_time = time.perf_counter()
-
-        # Publish event every 100ms
-        if (current_time - last_publish) >= 0.1:
-            event = source.publish_event()
-            print(f"[{current_time - start_time:.3f}s] Published: Event_{event['event_id']}")
-            last_publish = current_time
-
-            # Process event SYNCHRONOUSLY (BLOCKS!)
-            print(f"  → Processing Event_{event['event_id']} (500ms CPU work)...")
-            result = cpu_intensive_work(event["event_id"], duration_ms=500)
-            processed_events.append(result)
-            proc_time = result["processing_time_ms"]
-            print(f"  ✓ Completed Event_{event['event_id']} in {proc_time:.0f}ms")
-
-    # Results
-    print()
-    print(f"Total events published: {source.events_published}")
-    print(f"Total events processed: {len(processed_events)}")
-    print(f"Events MISSED: {source.events_published - len(processed_events)}")
-    print()
-
-    return {
-        "published": source.events_published,
-        "processed": len(processed_events),
-        "missed": source.events_published - len(processed_events),
-    }
-
-
-# ============================================================================
-# Scenario 2: ASYNC (Naive - DOESN'T Solve CPU Problem!)
-# ============================================================================
-
-
-async def scenario_2_async_naive():  # noqa: ANN201
-    """
-    Async event processing with CPU-bound work.
-
-    Expected: STILL misses events because CPU work blocks event loop!
-
-    Common misconception: "async solves everything"
-    Reality: async helps with I/O, NOT CPU work
-    """
-    print("=" * 80)
-    print("SCENARIO 2: ASYNC (Naive - Doesn't Help!)")
-    print("=" * 80)
-    print("Publishing events every 100ms (async)")
-    print("Processing takes 500ms (CPU-bound work - BLOCKS EVENT LOOP!)")
-    print("Expected: Events STILL missed (async doesn't help CPU work)")
-    print()
-
-    source = EventSource(interval_ms=100)
-    processed_events = []
-    event_queue = asyncio.Queue()
-
-    async def publisher() -> None:
-        """Publish events every 100ms."""
-        while True:
-            event = source.publish_event()
-            elapsed = time.perf_counter() - start_time
-            print(f"[{elapsed:.3f}s] Published: Event_{event['event_id']}")
-            await event_queue.put(event)
-            await asyncio.sleep(0.1)  # 100ms interval
-
-    async def worker() -> None:
-        """Process events from queue."""
-        while True:
-            event = await event_queue.get()
-            elapsed = time.perf_counter() - start_time
-            print(f"  [{elapsed:.3f}s] Processing Event_{event['event_id']} (500ms CPU work)...")
-
-            # ❌ THIS BLOCKS THE EVENT LOOP!
-            # Even though we're in async function, this is synchronous CPU work
-            result = cpu_intensive_work(event["event_id"], duration_ms=500)
-
-            processed_events.append(result)
-            elapsed = time.perf_counter() - start_time
-            print(f"  [{elapsed:.3f}s] ✓ Completed Event_{event['event_id']}")
-
-    # Start tasks
-    start_time = time.perf_counter()
-    publisher_task = asyncio.create_task(publisher())
-    worker_task = asyncio.create_task(worker())
-
-    # Run for 3 seconds
-    await asyncio.sleep(3.0)
-
-    # Stop tasks
-    publisher_task.cancel()
-    worker_task.cancel()
-
-    with contextlib.suppress(asyncio.CancelledError):
-        await publisher_task
-
-    with contextlib.suppress(asyncio.CancelledError):
-        await worker_task
-
-    # Results
-    print()
-    print(f"Total events published: {source.events_published}")
-    print(f"Total events processed: {len(processed_events)}")
-    print(f"Events in queue: {event_queue.qsize()}")
-    print(f"Events MISSED: {source.events_published - len(processed_events) - event_queue.qsize()}")
-    print()
-    print("⚠️ ASYNC DIDN'T HELP! CPU work still blocks event loop!")
-    print()
-
-    return {
-        "published": source.events_published,
-        "processed": len(processed_events),
-        "queued": event_queue.qsize(),
-        "missed": source.events_published - len(processed_events) - event_queue.qsize(),
-    }
-
-
-# ============================================================================
-# Scenario 3: ASYNC + ProcessPoolExecutor (REAL Solution)
-# ============================================================================
-
-
-async def scenario_3_async_processpool():  # noqa: ANN201
-    """
-    Async event processing with ProcessPoolExecutor.
-
-    Expected: NO missed events (true parallelism via separate processes).
-
-    This is the REAL solution for CPU-bound work in async Python.
-    """
-    print("=" * 80)
-    print("SCENARIO 3: ASYNC + ProcessPoolExecutor (TRUE Solution)")
-    print("=" * 80)
-    print("Publishing events every 100ms (async)")
-    print("Processing takes 500ms (CPU-bound work in SEPARATE PROCESS)")
-    print("Expected: NO missed events (true parallelism)")
-    print()
-
-    source = EventSource(interval_ms=100)
-    processed_events = []
-    event_queue = asyncio.Queue()
-
-    # Create process pool (separate Python processes - NO GIL!)
-    num_workers = 2  # 2 parallel processes
-    executor = ProcessPoolExecutor(max_workers=num_workers)
-
-    async def publisher() -> None:
-        """Publish events every 100ms."""
-        while True:
-            event = source.publish_event()
-            elapsed = time.perf_counter() - start_time
-            print(f"[{elapsed:.3f}s] Published: Event_{event['event_id']}")
-            await event_queue.put(event)
-            await asyncio.sleep(0.1)  # 100ms interval
-
-    async def worker(worker_id: int) -> None:
-        """Process events from queue using ProcessPoolExecutor."""
-        loop = asyncio.get_event_loop()
-
-        while True:
-            event = await event_queue.get()
-            elapsed = time.perf_counter() - start_time
-            print(
-                f"  [{elapsed:.3f}s] Worker-{worker_id}: Processing Event_{event['event_id']} "
-                f"(500ms CPU work in separate process)..."
-            )
-
-            # ✅ Run CPU work in separate process (NON-BLOCKING!)
-            result = await loop.run_in_executor(
-                executor,
-                cpu_intensive_work,
-                event["event_id"],
-                500,  # duration_ms
-            )
-
-            processed_events.append(result)
-            elapsed = time.perf_counter() - start_time
-            print(f"  [{elapsed:.3f}s] Worker-{worker_id}: ✓ Completed Event_{event['event_id']}")
-
-    # Start tasks
-    start_time = time.perf_counter()
-    publisher_task = asyncio.create_task(publisher())
-    worker_tasks = [asyncio.create_task(worker(i)) for i in range(num_workers)]
-
-    # Run for 3 seconds
-    await asyncio.sleep(3.0)
-
-    # Stop tasks
-    publisher_task.cancel()
-    for task in worker_tasks:
-        task.cancel()
-
-    with contextlib.suppress(asyncio.CancelledError):
-        await publisher_task
-
-    for task in worker_tasks:
-        with contextlib.suppress(asyncio.CancelledError):
-            await task
-
-    # Cleanup
-    executor.shutdown(wait=False)
-
-    # Results
-    print()
-    print(f"Total events published: {source.events_published}")
-    print(f"Total events processed: {len(processed_events)}")
-    print(f"Events in queue: {event_queue.qsize()}")
-    print(f"Events MISSED: {source.events_published - len(processed_events) - event_queue.qsize()}")
-    print()
-    print("✅ ProcessPoolExecutor WORKS! No events missed!")
-    print()
-
-    return {
-        "published": source.events_published,
-        "processed": len(processed_events),
-        "queued": event_queue.qsize(),
-        "missed": source.events_published - len(processed_events) - event_queue.qsize(),
-    }
-
-
-# ============================================================================
-# Main Test Runner
-# ============================================================================
-
-
-async def main() -> None:
-    """Run all scenarios and compare results."""
-    print()
-    print("=" * 80)
-    print("REAL Async Event Architecture - Proof of Concept")
-    print("=" * 80)
-    print()
-    print("This demonstrates the REAL problem with CPU-bound work in Python:")
-    print("- asyncio.sleep() is NOT the same as CPU work")
-    print("- CPU work BLOCKS the event loop")
-    print("- ProcessPoolExecutor provides TRUE parallelism")
-    print()
-
-    # Scenario 1: Synchronous (baseline)
-    result1 = scenario_1_synchronous()
-
-    # Scenario 2: Async (naive - doesn't help)
-    result2 = await scenario_2_async_naive()
-
-    # Scenario 3: Async + ProcessPool (real solution)
-    result3 = await scenario_3_async_processpool()
-
-    # Summary
-    print()
-    print("=" * 80)
-    print("SUMMARY - Events Published vs Processed")
-    print("=" * 80)
-    print()
-    print(
-        f"Scenario 1 (SYNC):           Published: {result1['published']:2d}, "
-        f"Processed: {result1['processed']:2d}, Missed: {result1['missed']:2d}"
-    )
-    print(
-        f"Scenario 2 (ASYNC naive):    Published: {result2['published']:2d}, "
-        f"Processed: {result2['processed']:2d}, Queued: {result2['queued']:2d}, "
-        f"Missed: {result2['missed']:2d}"
-    )
-    print(
-        f"Scenario 3 (ProcessPool):    Published: {result3['published']:2d}, "
-        f"Processed: {result3['processed']:2d}, Queued: {result3['queued']:2d}, "
-        f"Missed: {result3['missed']:2d}"
-    )
-    print()
-    print("=" * 80)
-    print("CONCLUSION")
-    print("=" * 80)
-    print()
-    print("1. Synchronous: Events missed during blocking CPU work ❌")
-    print("2. Async (naive): Async DOESN'T help with CPU work ❌")
-    print("3. ProcessPoolExecutor: TRUE parallelism solves it ✅")
-    print()
-    print("For SimpleTraderV3:")
-    print("- Use async for I/O (API calls, database, EventBus)")
-    print("- Use ProcessPoolExecutor for CPU work (signal detection, risk calc)")
-    print("- Combine both for complete solution")
-    print()
-
-
-if __name__ == "__main__":
-    # Set start method for multiprocessing (required on Windows)
-    multiprocessing.set_start_method("spawn", force=True)
-
-    asyncio.run(main())
