--- temp/artifacts_diff.txt
@@ -1,166 +1,0 @@
-﻿diff --git a/.st3/config/artifacts.yaml b/.st3/config/artifacts.yaml
-index 79654408..79557e54 100644
---- a/.st3/config/artifacts.yaml
-+++ b/.st3/config/artifacts.yaml
-@@ -49,23 +49,24 @@ artifact_types:
-       states: [CREATED]
-       initial_state: CREATED
-       valid_transitions: []
--  - type: code
--    type_id: adapter
--    name: "Adapter"
--    description: "Adapter - integrates external systems (implements Interface)"
--    output_type: "file"
--    scaffolder_class: "AdapterScaffolder"
--    scaffolder_module: "mcp_server.scaffolders.adapter_scaffolder"
--    template_path: null  # No template needed - adapter uses legacy scaffolder
--    fallback_template: null
--    name_suffix: "Adapter"
--    file_extension: ".py"
--    generate_test: true
--    # NOTE: NO base_path or test_base_path (D2 - use project_structure.yaml only)
--    state_machine:
--      states: [CREATED]
--      initial_state: CREATED
--      valid_transitions: []
-+  # DISABLED (issue #325 - no V2 pipeline support, no template):
-+  # - type: code
-+  #   type_id: adapter
-+  #   name: "Adapter"
-+  #   description: "Adapter - integrates external systems (implements Interface)"
-+  #   output_type: "file"
-+  #   scaffolder_class: "AdapterScaffolder"
-+  #   scaffolder_module: "mcp_server.scaffolders.adapter_scaffolder"
-+  #   template_path: null  # No template needed - adapter uses legacy scaffolder
-+  #   fallback_template: null
-+  #   name_suffix: "Adapter"
-+  #   file_extension: ".py"
-+  #   generate_test: true
-+  #   # NOTE: NO base_path or test_base_path (D2 - use project_structure.yaml only)
-+  #   state_machine:
-+  #     states: [CREATED]
-+  #     initial_state: CREATED
-+  #     valid_transitions: []
-   - type: code
-     type_id: tool
-     name: "MCP Tool"
-@@ -83,23 +84,24 @@ artifact_types:
-       states: [CREATED]
-       initial_state: CREATED
-       valid_transitions: []
--  - type: code
--    type_id: resource
--    name: "MCP Resource"
--    description: "MCP Resource - provides dynamic content"
--    output_type: "file"
--    scaffolder_class: "ResourceScaffolder"
--    scaffolder_module: "mcp_server.scaffolders.resource_scaffolder"
--    template_path: null  # No template needed - resource uses legacy scaffolder
--    fallback_template: null
--    name_suffix: null
--    file_extension: ".py"
--    generate_test: true
--    # NOTE: NO base_path or test_base_path (D2 - use project_structure.yaml only)
--    state_machine:
--      states: [CREATED]
--      initial_state: CREATED
--      valid_transitions: []
-+  # DISABLED (issue #325 - no V2 pipeline support, no template):
-+  # - type: code
-+  #   type_id: resource
-+  #   name: "MCP Resource"
-+  #   description: "MCP Resource - provides dynamic content"
-+  #   output_type: "file"
-+  #   scaffolder_class: "ResourceScaffolder"
-+  #   scaffolder_module: "mcp_server.scaffolders.resource_scaffolder"
-+  #   template_path: null  # No template needed - resource uses legacy scaffolder
-+  #   fallback_template: null
-+  #   name_suffix: null
-+  #   file_extension: ".py"
-+  #   generate_test: true
-+  #   # NOTE: NO base_path or test_base_path (D2 - use project_structure.yaml only)
-+  #   state_machine:
-+  #     states: [CREATED]
-+  #     initial_state: CREATED
-+  #     valid_transitions: []
-   - type: code
-     type_id: schema
-     name: "Pydantic Schema"
-@@ -117,23 +119,24 @@ artifact_types:
-       states: [CREATED]
-       initial_state: CREATED
-       valid_transitions: []
--  - type: code
--    type_id: interface
--    name: "Interface"
--    description: "Interface - abstract protocol definition"
--    output_type: "file"
--    scaffolder_class: "InterfaceScaffolder"
--    scaffolder_module: "mcp_server.scaffolders.interface_scaffolder"
--    template_path: null  # No template needed - interface uses legacy scaffolder
--    fallback_template: null
--    name_suffix: null
--    file_extension: ".py"
--    generate_test: true
--    # NOTE: NO base_path or test_base_path (D2 - use project_structure.yaml only)
--    state_machine:
--      states: [CREATED]
--      initial_state: CREATED
--      valid_transitions: []
-+  # DISABLED (issue #325 - no V2 pipeline support, no template):
-+  # - type: code
-+  #   type_id: interface
-+  #   name: "Interface"
-+  #   description: "Interface - abstract protocol definition"
-+  #   output_type: "file"
-+  #   scaffolder_class: "InterfaceScaffolder"
-+  #   scaffolder_module: "mcp_server.scaffolders.interface_scaffolder"
-+  #   template_path: null  # No template needed - interface uses legacy scaffolder
-+  #   fallback_template: null
-+  #   name_suffix: null
-+  #   file_extension: ".py"
-+  #   generate_test: true
-+  #   # NOTE: NO base_path or test_base_path (D2 - use project_structure.yaml only)
-+  #   state_machine:
-+  #     states: [CREATED]
-+  #     initial_state: CREATED
-+  #     valid_transitions: []
-   - type: code
-     type_id: service
-     name: "Service"
-@@ -288,20 +291,21 @@ artifact_types:
-           to: [APPROVED, DRAFT]
-         - from: APPROVED
-           to: [DEFINITIVE]
--  - type: doc
--    type_id: tracking
--    name: "Tracking Document"
--    description: "Living document for tracking project status"
--    output_type: "file"
--    template_path: "docs/tracking.md.jinja2"
--    fallback_template: null
--    name_suffix: null
--    file_extension: ".md"
--    generate_test: false
--    state_machine:
--      states: [LIVING DOCUMENT]
--      initial_state: LIVING DOCUMENT
--      valid_transitions: []
-+  # DISABLED (issue #325 - no V2 pipeline support, stale template path, superseded by commit/pr/issue):
-+  # - type: doc
-+  #   type_id: tracking
-+  #   name: "Tracking Document"
-+  #   description: "Living document for tracking project status"
-+  #   output_type: "file"
-+  #   template_path: "docs/tracking.md.jinja2"
-+  #   fallback_template: null
-+  #   name_suffix: null
-+  #   file_extension: ".md"
-+  #   generate_test: false
-+  #   state_machine:
-+  #     states: [LIVING DOCUMENT]
-+  #     initial_state: LIVING DOCUMENT
-+  #     valid_transitions: []
-   - type: doc
-     type_id: reference
-     name: "Reference Document"
