--- temp/branch_py_files.txt
@@ -1,105 +1,0 @@
-backend/dtos/validation_fixture_gate4.py
-mcp_server/config/quality_config.py
-mcp_server/managers/qa_manager.py
-mcp_server/server.py
-mcp_server/tools/quality_tools.py
-mcp_server/tools/test_tools.py
-mcp_server/utils/__init__.py
-mcp_server/utils/path_resolver.py
-tests/mcp_server/integration/test_qa.py
-tests/mcp_server/unit/adapters/__init__.py
-tests/mcp_server/unit/adapters/test_git_adapter.py
-tests/mcp_server/unit/adapters/test_github_adapter.py
-tests/mcp_server/unit/config/test_json_violations_parsing.py
-tests/mcp_server/unit/config/test_label_config.py
-tests/mcp_server/unit/config/test_label_startup.py
-tests/mcp_server/unit/config/test_labelconfig_singleton_bug.py
-tests/mcp_server/unit/config/test_quality_config.py
-tests/mcp_server/unit/config/test_settings.py
-tests/mcp_server/unit/config/test_text_violations_parsing.py
-tests/mcp_server/unit/config/test_text_violations_parsing_defaults_validator.py
-tests/mcp_server/unit/config/test_violation_dto.py
-tests/mcp_server/unit/config/test_workflow_config.py
-tests/mcp_server/unit/conftest.py
-tests/mcp_server/unit/core/test_error_handling.py
-tests/mcp_server/unit/core/test_logging.py
-tests/mcp_server/unit/core/test_mcp_exceptions.py
-tests/mcp_server/unit/integration/__init__.py
-tests/mcp_server/unit/integration/test_all_tools.py
-tests/mcp_server/unit/integration/test_fs_status.py
-tests/mcp_server/unit/integration/test_git.py
-tests/mcp_server/unit/integration/test_github.py
-tests/mcp_server/unit/managers/test_auto_scope_resolution.py
-tests/mcp_server/unit/managers/test_autofix_propagation.py
-tests/mcp_server/unit/managers/test_baseline_advance.py
-tests/mcp_server/unit/managers/test_compact_payload_builder.py
-tests/mcp_server/unit/managers/test_deliverable_checker.py
-tests/mcp_server/unit/managers/test_execute_gate_dispatch.py
-tests/mcp_server/unit/managers/test_extract_violations_array.py
-tests/mcp_server/unit/managers/test_feature_flag_v2.py
-tests/mcp_server/unit/managers/test_files_for_gate.py
-tests/mcp_server/unit/managers/test_filter_files_removed.py
-tests/mcp_server/unit/managers/test_git_manager.py
-tests/mcp_server/unit/managers/test_github_manager.py
-tests/mcp_server/unit/managers/test_legacy_parsers_removed.py
-tests/mcp_server/unit/managers/test_parse_json_violations.py
-tests/mcp_server/unit/managers/test_parse_json_violations_nested.py
-tests/mcp_server/unit/managers/test_parse_json_violations_options.py
-tests/mcp_server/unit/managers/test_parse_text_violations.py
-tests/mcp_server/unit/managers/test_parse_text_violations_defaults.py
-tests/mcp_server/unit/managers/test_phase_state_engine_c2.py
-tests/mcp_server/unit/managers/test_phase_state_engine_c3.py
-tests/mcp_server/unit/managers/test_phase_state_engine_parent_branch.py
-tests/mcp_server/unit/managers/test_phase_state_engine_persistence.py
-tests/mcp_server/unit/managers/test_phase_state_engine_recovery.py
-tests/mcp_server/unit/managers/test_phase_state_engine_workflow.py
-tests/mcp_server/unit/managers/test_project_manager.py
-tests/mcp_server/unit/managers/test_pyright_severity_mapping.py
-tests/mcp_server/unit/managers/test_pytest_helpers_removed.py
-tests/mcp_server/unit/managers/test_qa_manager.py
-tests/mcp_server/unit/managers/test_scope_resolution.py
-tests/mcp_server/unit/managers/test_skip_reason_unified.py
-tests/mcp_server/unit/managers/test_summary_c39.py
-tests/mcp_server/unit/managers/test_summary_line_formatter.py
-tests/mcp_server/unit/managers/test_violation_path_normalization.py
-tests/mcp_server/unit/resources/__init__.py
-tests/mcp_server/unit/resources/test_base.py
-tests/mcp_server/unit/resources/test_resources_base.py
-tests/mcp_server/unit/resources/test_standards.py
-tests/mcp_server/unit/schemas/__init__.py
-tests/mcp_server/unit/schemas/test_code_artifact_parity.py
-tests/mcp_server/unit/schemas/test_code_artifact_schemas.py
-tests/mcp_server/unit/schemas/test_code_artifact_v2_parity.py
-tests/mcp_server/unit/schemas/test_doc_artifact_v2_parity.py
-tests/mcp_server/unit/schemas/test_dto_schemas.py
-tests/mcp_server/unit/schemas/test_lifecycle.py
-tests/mcp_server/unit/schemas/test_tracking_artifact_v2_parity.py
-tests/mcp_server/unit/test_cli.py
-tests/mcp_server/unit/test_dto_parity.py
-tests/mcp_server/unit/test_server.py
-tests/mcp_server/unit/tools/test_base_tool_error_handling.py
-tests/mcp_server/unit/tools/test_dev_tools.py
-tests/mcp_server/unit/tools/test_force_phase_transition_tool.py
-tests/mcp_server/unit/tools/test_git_checkout_async.py
-tests/mcp_server/unit/tools/test_git_checkout_state_sync.py
-tests/mcp_server/unit/tools/test_git_fetch_tool.py
-tests/mcp_server/unit/tools/test_git_fetch_tool_behavior.py
-tests/mcp_server/unit/tools/test_git_pull_tool.py
-tests/mcp_server/unit/tools/test_git_pull_tool_behavior.py
-tests/mcp_server/unit/tools/test_initialize_project_tool.py
-tests/mcp_server/unit/tools/test_label_tools_integration.py
-tests/mcp_server/unit/tools/test_project_tools.py
-tests/mcp_server/unit/tools/test_quality_tools.py
-tests/mcp_server/unit/tools/test_test_tools.py
-tests/mcp_server/unit/tools/test_tool_result_contract.py
-tests/mcp_server/unit/tools/test_transition_phase_tool.py
-tests/mcp_server/unit/utils/__init__.py
-tests/mcp_server/unit/utils/test_path_resolver.py
-tests/mcp_server/unit/validation/test_markdown_validator.py
-tests/mcp_server/unit/validation/test_python_validator.py
-tests/mcp_server/unit/validation/test_template_analyzer.py
-tests/mcp_server/unit/validation/test_validation_service.py
-tests/mcp_server/unit/validation/test_validator_registry.py
-tests/mcp_server/validation_fixtures/__init__.py
-tests/mcp_server/validation_fixtures/gate0_format_violation.py
-tests/mcp_server/validation_fixtures/violations.py
