--- temp/ps_diff.txt
@@ -1,72 +1,0 @@
-﻿diff --git a/.st3/config/project_structure.yaml b/.st3/config/project_structure.yaml
-index 209addcf..965f7e2c 100644
---- a/.st3/config/project_structure.yaml
-+++ b/.st3/config/project_structure.yaml
-@@ -10,8 +10,6 @@ directories:
-     allowed_artifact_types:
-       - dto
-       - worker
--      - adapter
--      - interface
-       - service
-     allowed_extensions:
-       - ".py"
-@@ -35,14 +33,12 @@ directories:
-   backend/adapters:
-     parent: backend
-     description: "External system adapters"
--    allowed_artifact_types:
--      - adapter
-+    allowed_artifact_types: []  # DISABLED (issue #325)
- 
-   backend/interfaces:
-     parent: backend
-     description: "Abstract protocol definitions"
--    allowed_artifact_types:
--      - interface
-+    allowed_artifact_types: []  # DISABLED (issue #325)
- 
-   backend/services:
-     parent: backend
-@@ -55,7 +51,6 @@ directories:
-     description: "MCP platform code"
-     allowed_artifact_types:
-       - tool
--      - resource
-       - schema
-       - generic
-     allowed_extensions:
-@@ -73,8 +68,7 @@ directories:
-   mcp_server/resources:
-     parent: mcp_server
-     description: "MCP resource providers"
--    allowed_artifact_types:
--      - resource
-+    allowed_artifact_types: []  # DISABLED (issue #325)
- 
-   mcp_server/schemas:
-     parent: mcp_server
-@@ -135,7 +129,6 @@ directories:
-       - planning
-       - design
-       - architecture
--      - tracking
-       - reference
-       - generic
-     allowed_extensions:
-@@ -157,7 +150,6 @@ directories:
-       - research
-       - planning
-       - design
--      - tracking
- 
-   docs/reference:
-     parent: docs
-@@ -184,7 +176,6 @@ directories:
-     parent: docs
-     description: "Implementation tracking documents"
-     allowed_artifact_types:
--      - tracking
-       - reference
- 
-   .st3:
