--- temp/server_diff.txt
@@ -1,53 +1,0 @@
-﻿diff --git a/mcp_server/server.py b/mcp_server/server.py
-index 52e160ac..61f5a496 100644
---- a/mcp_server/server.py
-+++ b/mcp_server/server.py
-@@ -452,13 +452,13 @@ class MCPServer:
- 
-     def _validate_tool_arguments(
-         self, tool: BaseTool, arguments: dict[str, Any] | None, call_id: str, name: str
--    ) -> BaseModel | dict[str, Any] | list[TextContent | ImageContent | EmbeddedResource]:
-+    ) -> BaseModel | dict[str, Any] | ToolResult:
-         """Validate tool arguments against args_model.
- 
-         Returns:
-             - Validated BaseModel instance if validation succeeds
-             - Raw arguments dict if no args_model
--            - List of content with error if validation fails
-+            - ToolResult with is_error=True if validation fails
-         """
-         if not getattr(tool, "args_model", None):
-             return arguments or {}
-@@ -500,7 +500,20 @@ class MCPServer:
-                 },
-             )
-             error_details = str(validation_error)
--            return [TextContent(type="text", text=f"Invalid input for {name}: {error_details}")]
-+            return ToolResult(
-+                content=[
-+                    {"type": "text", "text": f"Invalid input for {name}: {error_details}"},
-+                    {
-+                        "type": "resource",
-+                        "resource": {
-+                            "uri": "schema://validation",
-+                            "mimeType": "application/json",
-+                            "text": json.dumps(tool.input_schema),
-+                        },
-+                    },
-+                ],
-+                is_error=True,
-+            )
- 
-     def _convert_tool_result_to_content(
-         self, result: ToolResult
-@@ -624,8 +637,8 @@ class MCPServer:
-                         # Validate arguments
-                         validated = self._validate_tool_arguments(tool, arguments, call_id, name)
-                         # Early return if validation failed
--                        if isinstance(validated, list):
--                            return validated
-+                        if isinstance(validated, ToolResult):
-+                            return self._convert_tool_result_to_mcp_result(validated)
- 
-                         note_context = NoteContext()
-
