--- temp/test-design-doc.md
@@ -1,67 +1,0 @@
-<!-- temp\test-design-doc.md -->
-<!-- template=design version=5827e841 created=2026-06-05T13:41Z updated= -->
-# Test Design Document
-
-**Status:** DRAFT  
-**Version:** 1.0  
-**Last Updated:** 2026-06-05
-
----
-
-## Purpose
-
-Smoke test for the design template after structural field migration
-
-## Scope
-
-**In Scope:**
-design.md.jinja2 output validation
-
-**Out of Scope:**
-Other template types
-
----
-
-## 1. Context & Requirements
-
-### 1.1. Problem Statement
-
-Verify that the design template scaffolds correctly after the DocArtifactContext structural refactor
-
-### 1.2. Requirements
-
-**Functional:**
-- [ ] Template renders all sections correctly
-- [ ] Structural fields (status, version, last_updated) appear in header
-
-**Non-Functional:**
-- [ ] Output is valid Markdown
-
-### 1.3. Constraints
-
-None
----
-
-## 2. Design Options
----
-
-## 3. Chosen Design
-
-**Decision:** Use the updated design template with required structural fields in DocArtifactContext
-
-**Rationale:** Validates Finding 8 fix from C_286.6
-
-### 3.1. Key Design Decisions
-
-| Decision | Rationale |
-|----------|-----------|
-
-## Related Documentation
-None
----
-
-## Version History
-
-| Version | Date | Author | Changes |
-|---------|------|--------|---------|
-| 1.0 | 2026-06-05 | Agent | Initial draft |
